+++ handlers/product-hub/helpers/check-if-all-handlers-exist.ts
