+++ handlers/product-hub/helpers/index.ts
@@ -1,8 +1,8 @@
-export { checkIfAllHandlersExist } from 'handlers/product-hub/helpers/checkIfAllHandlersExist'
-export { filterTableData } from 'handlers/product-hub/helpers/filterTableData'
-export { findByIlk } from 'handlers/product-hub/helpers/findByIlk'
-export { findByTokenPair } from 'handlers/product-hub/helpers/findByTokenPair'
-export { getProtocolProducts } from 'handlers/product-hub/helpers/getProtocolProducts'
-export { getTokenGroup } from 'handlers/product-hub/helpers/getTokenGroup'
+export { checkIfAllHandlersExist } from 'handlers/product-hub/helpers/check-if-all-handlers-exist'
+export { filterTableData } from 'handlers/product-hub/helpers/filter-table-data'
+export { findByIlk } from 'handlers/product-hub/helpers/find-by-ilk'
+export { findByTokenPair } from 'handlers/product-hub/helpers/find-by-token-pair'
+export { getProtocolProducts } from 'handlers/product-hub/helpers/get-protocol-products'
+export { getTokenGroup } from 'handlers/product-hub/helpers/get-token-group'
 export { aaveLikeAprToApy } from 'handlers/product-hub/helpers/aave-like-apr-to-apy'
 export * from 'handlers/product-hub/helpers/constants'
